requirements.txt
pip>=19.0.3 
setuptools==40.4.3 
requests==2.20.0 
robobrowser==0.5.3 
lxml==4.2.5 
pytest==3.9.1 
simplejson==3.16.0 
vcrpy==2.0.1 
flake8==3.6.0 
jupyter==1.0.0
